requirements.txt
# Core
arxiv>=2.1.0
pydantic>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
httpx>=0.27.0

# LLM Providers (install as needed)
anthropic>=0.40.0      # Claude
openai>=1.0.0          # OpenAI
ollama>=0.4.0          # Ollama local models
